Move Detect_feature_edges to PMP
