+++ Mesh_3/include/CGAL/Polyhedral_mesh_domain_with_features_3.h
@@ -39,7 +39,7 @@
 
 #include <CGAL/Mesh_3/Detect_polylines_in_polyhedra.h>
 #include <CGAL/Mesh_3/Polyline_with_context.h>
-#include <CGAL/Mesh_3/Detect_features_in_polyhedra.h>
+#include <CGAL/Polygon_mesh_processing/Detect_features_in_polyhedra.h>
 
 #include <CGAL/enum.h>
 #include <CGAL/IO/Polyhedron_iostream.h>
@@ -421,7 +421,7 @@ detect_features(FT angle_in_degree, std::vector<Polyhedron>& poly)
   P2vmap p2vmap;
 
   typedef typename boost::property_map<Polyhedron,face_patch_id_t>::type Face_patch_id_pmap;
-  CGAL::Mesh_3::Detect_features_in_polyhedra<Polyhedron,Face_patch_id_pmap> detect_features(get(face_patch_id,poly.front()));
+  CGAL::Polygon_mesh_processing::Detect_features_in_polyhedra<Polyhedron,Face_patch_id_pmap> detect_features(get(face_patch_id,poly.front()));
   BOOST_FOREACH(Polyhedron& p, poly)
   {
     initialize_ts(p);
